requirements.txt
langchain
openai
huggingface_hub
python-dotenv
streamlit
unstructured
pinecone-client
tiktoken
pandas
numpy
streamlit
langchain-google-genai